Reworking the message system
We are now waiting for tcp messages to reach the destination to send a new one

+++ Assets/_Scripts/GameManager.cs
@@ -30,7 +30,7 @@ public class GameManager : MonoBehaviour {
             await Task.Delay(1000);
         }
 
-        ConnectionManager.SendMessage(MessageType.UpdateGamePhase, GamePhase.VotingForDestination);
+        ConnectionManager.SendMessage(MessageType.UpdateGamePhase, (int)GamePhase.VotingForDestination);
 
         await Task.Delay(1000);
